+++ WheelWizard/Features/WiiManagement/WiiManagementExtensions.cs
@@ -1,4 +1,5 @@
-﻿using WheelWizard.WiiManagement.Domain.Mii;
+﻿using WheelWizard.Services.Settings;
+using WheelWizard.WiiManagement.Domain.Mii;
 
 namespace WheelWizard.WiiManagement;
 
@@ -46,5 +47,26 @@ public static class WiiManagementExtensions
 
         cloneResult.Value.MiiId = miiId; // IMPORTANT: Also make sure that the clone has the same id as the OG, it's a clone after all.
         return cloneResult.Value;
+    }
+
+    public static bool IsGlobal(this Mii self)
+    {
+        // If it has blue pants, then its definitely global
+        if ((self.MiiId1 >> 5) == 0b110)
+            return true;
+
+        // But it can also be global if the mac address is not the same as your own address
+        var macAddressString = (string)SettingsManager.MACADDRESS.Get();
+        var macParts = macAddressString.Split(':');
+        var macBytes = new byte[6];
+        for (var i = 0; i < 6; i++)
+            macBytes[i] = byte.Parse(macParts[i], System.Globalization.NumberStyles.HexNumber);
+        var systemId0 = (byte)((macBytes[0] + macBytes[1] + macBytes[2]) & 0xFF);
+        return (
+            self?.SystemId0 != systemId0
+            || self?.SystemId1 != macBytes[3]
+            || self?.SystemId2 != macBytes[4]
+            || self?.SystemId3 != macBytes[5]
+        );
     }
 }
